+++ src/util.rs
@@ -1,10 +1,8 @@
-use std::sync::atomic::AtomicU32;
+use std::sync::atomic::AtomicU64;
 
-static ID: AtomicU32 = AtomicU32::new(0);
+static ID: AtomicU64 = AtomicU64::new(0);
 
 /// Retrieve a new, globally-unique ID.
-pub fn new_id() -> u32 {
+pub fn new_id() -> u64 {
     ID.fetch_add(1, std::sync::atomic::Ordering::Relaxed)
 }
-
-
